requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
streamlit>=1.36.0
pandas>=2.2.0
geopy>=2.4.0
thefuzz>=0.22.0
python-Levenshtein>=0.25.0
